Move 'FeedViewAdaptor' to its own file
